backend-python/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
bitarray==3.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==1.0.2
click==8.1.8
cytoolz==1.1.0
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.1
fastapi==0.115.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
hexbytes==0.3.1
httptools==0.7.1
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lru-dict==1.2.0
multidict==6.7.0
parsimonious==0.10.0
pillow==10.4.0
propcache==0.4.1
protobuf==6.33.1
psycopg2-binary==2.9.11
py-ecc==8.0.0
pycryptodome==3.23.0
pydantic==2.9.2
pydantic-settings==2.11.0
pydantic_core==2.23.4
pypng==0.20220715.0
python-dotenv==1.0.1
pyunormalize==17.0.0
PyYAML==6.0.3
qrcode==7.4.2
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
rpds-py==0.27.1
SQLAlchemy==2.0.44
starlette==0.38.6
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
web3==6.20.1
websockets==15.0.1
yarl==1.22.0
